Move hoconJsonToNetworkDefinition to modal component.
